pyproject.toml
[tool.poetry]
name = "oa4-price-formation"
version = "0.1.0"
description = ""
authors = ["Khitrov <Hitrov.95@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
coverage = "^7.5.3"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
flake8 = "^7.0.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
